Some rtkkw.py modifications
The former <br> tag was creating an unwanted line break after the last entry. To fix this, and make styling easier, I added the items inside a <p> instead. 

Added a `kw` class to the <p> to make the line easier to style as a whole.

Added a separator entry for the menu so so the user doesn't have to change the main file if they wish to alter it.

Added a `kw-separator` class to the separator for extra nitpicky users.


Possible bugfix?
The Keyword link was redirecting to the  VocabUrl instead of the kanjiUrl, given that there are 2 separate <a> entries, I assumed it was a bug.
  I corrected it to link to the kanjiUrl on the kanji entry and the vocabUrl on the vocab entry of the combined field.