config.json addition/mod
Added the previously mentioned separator user config.

Changed rtkDeck and rtkModel to reflect the author's shared deck page in a way the user doesn't have to think about those settings.

Changed "Reading" field to "Expression" due to fears of users trying to use their furigana/kana reading fields of the sentence decks and then having problems.
- It thought it was easier to rename than to explain why it was named "reading" and not "sentence" or "expression" which are more common in shared decks.

But if you think it will interfere with the compatibility of current users, do change it back to reading.

Also changed the default kanjiUrl back to kanjikoohi as it was on the previous version.